[feat] header 1:1채팅방 목록 가져오기 - 채팅방 info 불러오기 api, useQuery
