+++ app/_api/queryKeys.ts
@@ -38,3 +38,5 @@ export const QEURY_KEY_COMMUNITY_COMMENTS_LIST = `communityCommentsList`;
 export const QUERY_KEY_MESSAGES_LIST = `messages_list`;
 export const QUERY_KEY_PRIVATE_ROOM_IDS = `private_room_ids`;
 export const QUERY_KEY_PRIVATE_CHATS_LIST = `private_chats_list`;
+export const QUERY_KEY_MY_PRIVATE_ROOMS_IDS = `my_private_room_ids`;
+export const QUERY_KEY_ACTION_TITLES_URLS = `action_titles_urls`;
